+++ test/e2e/multi-part.ts
@@ -1,0 +1,35 @@
+import Koa = require('koa');
+import Router = require('koa-tree-router');
+import Busboy = require('busboy');
+
+export interface MultipartResult {
+  fields: { [key: string]: any };
+  files: { [key: string]: number };
+}
+
+export function parse_multipart(
+  ctx: Koa.ParameterizedContext<any, Router.IRouterParamContext<any, any>, any>,
+): Promise<MultipartResult> {
+  return new Promise(resolve => {
+    const busboy = new Busboy({ headers: ctx.headers });
+    const output: MultipartResult = { fields: {}, files: {} };
+
+    busboy.on('file', (fieldName, file) => {
+      let size = 0;
+      file.on('data', data => {
+        size += data.length;
+      });
+      file.on('end', () => {
+        output.files[fieldName] = size;
+      });
+    });
+    busboy.on('field', (fieldName, val) => {
+      output.fields[fieldName] = val;
+    });
+    busboy.on('finish', function() {
+      resolve(output);
+    });
+
+    ctx.req.pipe(busboy);
+  });
+}
